Refactor routing logic into a new class
